Text tool setting font as f32

+++ editor/src/messages/portfolio/document/node_graph/graph_operation_message.rs
@@ -92,7 +92,7 @@ pub enum GraphOperationMessage {
 		id: NodeId,
 		text: String,
 		font: Font,
-		size: f64,
+		size: f32,
 		parent: LayerNodeIdentifier,
 		insert_index: isize,
 	},
